Got rid of publishUI and moved its code into modals.js and main.js.
There wasn't a good conceptual or code separation between publishUI
and main/modals so I'm just killing it for now. Things are a bit more
confusing now but I think this will make it easier to refactor things
to be simpler in the long-term.
